add a "cast" to int in a double value to solve the bug

+++ AbilityScoreCalculator.cs
@@ -18,7 +18,7 @@ namespace AbilityScoreCalculator
         double divided = RollResult / DivideBy;
 
         // Add AddAmount to the result of that division
-        int added = AddAmount += divided;
+        int added = AddAmount += (int) divided;
 
         // If the result is too small, use Minimum
         if (added < Minimum)
